feat: styled Post
