+++ posts/1.mdx
@@ -5,82 +5,72 @@ date: 2020-07-18T20:11:24.568Z
 description: This is a test blog post.
 tags:
   - test
-banner: placeholder-6-.png
+thumbnail: death-note-L.jpg
+slug: test-post-one
 ---
-![Test image](placeholder-6-.png "Gravity Headers")
 
-## Background
+## Using GitHub Pages to deploy your GitHub Project
 
-There are a couple of issues that users face while integrating or working with Netlify CMS and while the community is doing their best fixing them, here is a list of known issues and their solutions.
+GitHub Pages is a static site hosting service that takes HTML, CSS, and JavaScript files straight from a repository on GitHub, optionally runs the files through a build process, and publishes a website.
 
-## API Errors
+It was created by GitHub to allow developers to host their projects directly from a GitHub repository without having to pay for domain cost.
 
-A list of error messages and their possible cause(s) to provide more context while debugging.
+You can publish a website about to `https://<username>.github.io` and your project to `https://<username>.github.io/{your-project-name}`.
 
-#### 1. Error Message.
+The purpose of this article is to guide you through publish your project with gh-pages whether it is a plain HTML, CSS & JavaScript project or it is built with React xD.
 
-```
-API_ERROR: {"message":"404 File Not Found"}
-```
 
-#### Possible Causes.
+### Plain HTML, CSS & JS
+1. If your website contains images make sure that they are not stored in a media folder as gh-pages doesn't parse relative imports well. You may have to upload your images to a service like Cloudinary.
 
-* The collection folder or file property is missing a value (a dedicated file or folder, can't be blank or root). The file or folder should be relative to root (so `src/posts` for example).
-* There are subfolders within the collection folder.
+2. With that out of the way, navigate to your repository settings and locating the GitHub Pages section
+![image](https://i.imgur.com/5ZlMuH7.png)
 
-#### 2. Error Message
+3. Select the master branch from the drop down menu and this triggers a refresh.
+![image](https://i.imgur.com/PDUMDMZ.png)
 
-```
-API_ERROR: Not Found
-```
+4. When the page is done refreshing, navigate back to the GitHub Pages section and it should be active with the URL to the live project active.
+![image](https://i.imgur.com/T7EH5CC.png)
 
-#### Possible Cause
+Note that because of this setting each commit you make to the master branch updates the live project too.
 
-* The media library folder is missing, or the value of the media library property within config.yml isn't correct.
+### React Project.
+1. Install gh-pages by running
 
-#### 3. Error Message
+`npm install gh-pages` or `yarn add gh-pages`
 
-```
-API_ERROR: Not Found
-```
+This packages creates a `gh-pages` branch that would hold the build files for your project.
 
-#### Possible Cause
-
-* The org account needs to grant access to your Netlify CMS instance. [More details](https://help.github.com/en/github/setting-up-and-managing-your-github-user-account/requesting-organization-approval-for-oauth-apps).
-
-#### 4. Error Message
-
-```
-Failed to load settings from /.netlify/identity
+2. In your package.json file
+add the following line 
+``` 
+"homepage": "https://<username>.github.io/<repo-name>",
 ```
 
-#### Possible Cause
+Note that the repo name can be anything not necessarily generic.
 
-* config.yml has git-gateway as a backend on a site that isn't hosted on Netlify.
-
-## Other Issues
-
-#### 1. Invited Users Email not Confirmed.
-
-If you’ve invited some users and opened the invitation link from the generated emails but as they try to log in they get the following error message:
-
+3. Also update your scripts to contain 
 ```
-Email not confirmed
+  "scripts": {
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build",
+    ....
+   },
 ```
 
-#### Possible Solutions.
+The predeploy command generates a build that is ready to be published while the deploy command generates a build then deploys it
+4. When you are ready run 
+```
+npm run deploy
+```
 
-* Add the identity widget to your homepage `<head>`. The upside of this method is that it’s fast and easy, no other configuration needed. The downside is that all of your users are downloading the widget while they might not need it. That’s kb’s wasted. Unless you’re using Netlify Identity for something else than the authentication of CMS users, utilize the next solution.
-* Change the invitation emails! This takes a bit more time, but it’s probably the cleanest solution. You already have the identity widget loaded at your `admin/` page, so you might as well have your users confirm their email there. Learn how to update your invitation emails [here](https://docs.netlify.com/visitor-access/identity/identity-generated-emails/#email-templates).
+Wait for it to finish deploying the site and check the URL you set as homepage in your package.json to view it.
 
-Credits to [@tomrutgers](https://github.com/tomrutgers). Check here for [more details](https://community.netlify.com/t/common-issue-netlify-cms-git-gateway-email-not-confirmed/10690).
-
-#### 2. Invitation Email not Received.
-
-If you deployed Netlify CMS to Netlify using a template and didn't receive an invitation email after deploying the site, check your email settings to make sure that your email provider doesn't block emails from no-reply@netlify.com.
-
-#### 3. Mobile responsiveness.
-
-To make Netlify CMS responsive for usage on mobile, [here](https://gist.github.com/lilpolymath/51082130d316df3a938648569077c933) is a temporary CSS solution to include to your `admin/index.html` below your closing `</body>` tag so it can overwrite netlify-generated css.
-
-Credits to [truongoi](https://github.com/truongoi).
+If you get the error 
+`git remote add origin https://github.com/lilpolymath/test-repo.git
+git push -u origin master`
+run 
+```
+rm -rf node_modules/gh-pages/.cache
+```
+that clears the cache for you and rerun the deploy command.
