+++ posts/2.mdx
@@ -5,7 +5,8 @@ date: 2020-07-16T20:11:24.568Z
 description: This is a test blog post.
 tags:
   - test
-banner: placeholder-6-.png
+thumbnail: death-note-L.jpg
+slug: test-post-two
 ---
 ![Test image](placeholder-6-.png "Gravity Headers")
