+++ posts/3.mdx
@@ -5,7 +5,8 @@ date: 2020-07-20T20:11:24.568Z
 description: This is a test blog post.
 tags:
   - test
-banner: placeholder-6-.png
+thumbnail: death-note-L.jpg
+slug: test-post-three
 ---
 
 How to add tracked files to .gitignore
