+++ src/pages/Post.jsx
@@ -1,56 +1,51 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
+import { Redirect } from 'react-router-dom';
+import { animated } from 'react-spring/renderprops';
 import Markdown from 'react-markdown';
+
 import postlist from '../posts.json';
+import styles from './Post.module.css';
 
-import './components.css';
+const Post = ({ style, mouseEnter, mouseLeave, ...props }) => {
+  const id = props.match.params.id;
 
-const PostList = () => {
-  const excerptList = postlist.map(post => {
-    return (
-      post.content
-        .split(' ')
-        .slice(0, 20)
-        .join(' ') + '...'
-    );
+  const slug = id;
+  if (!slug) {
+    return <Redirect to='/404' />;
+  }
+
+  const fetchedPost = {};
+  let postExists = false;
+  postlist.forEach((post, i) => {
+    if (slug === post.slug) {
+      fetchedPost.title = post.title ? post.title : 'No title given';
+      fetchedPost.date = post.date ? post.date : 'No date given';
+      fetchedPost.author = post.author ? post.author : 'No author given';
+      fetchedPost.content = post.content ? post.content : 'No content given';
+      fetchedPost.thumbnail = post.thumbnail
+        ? post.thumbnail
+        : 'No thumbnail set';
+      postExists = true;
+    }
   });
+
+  if (postExists === false) {
+    return <Redirect to='/404' />;
+  }
   return (
-    <div className='postlist'>
-      <h1 className='title'>All Posts</h1>
-      {postlist.length &&
-        postlist.map((post, i) => {
-          return (
-            <div key={i} className='post-card'>
-              <div className='img-container'>
-                {post.thumbnail && (
-                  <img
-                    className='thumbnail'
-                    width={80}
-                    src={post.thumbnail}
-                    alt=''
-                  />
-                )}
-                <h2 className='post-title'>
-                  <Link className='links' to={`/post/${post.id}`}>
-                    {post.title}
-                  </Link>
-                </h2>
-              </div>
-              <small>
-                Published on {post.date} by {post.author}
-              </small>
-              <hr />
-              <Markdown source={excerptList[i]} escapeHtml={false} />
-              <small>
-                <Link className='links' to={`/post/${post.id}`}>
-                  Read more
-                </Link>
-              </small>
-            </div>
-          );
-        })}
-    </div>
+    <animated.div style={style} className={styles.post}>
+      <img
+        src={fetchedPost.thumbnail}
+        className={styles.featured_image}
+        alt='thumbnail'
+      />
+      <h2 className={styles.title}>{fetchedPost.title}</h2>
+      {/* <small>Published on {fetchedPost.date} by Ayobami Favour</small> */}
+      <div className={styles.post_content}>
+        <Markdown source={fetchedPost.content} escapeHtml={false} />
+      </div>
+    </animated.div>
   );
 };
 
-export default PostList;
+export default Post;
