+++ src/pages/components.css
@@ -1,40 +1,8 @@
-h1 {
-  font-family: 'Raleway';
-  font-weight: 500;
-  margin: 0;
-}
-
 .links {
   text-decoration: none;
   color: darkblue;
 }
 
-.layout-body {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  min-height: 90vh;
-  padding: 20px;
-}
-
-.header {
-  background: navy;
-  color: lightblue;
-  padding: 10px 20px;
-}
-
-.navbar {
-  display: flex;
-  justify-content: space-around;
-}
-
-.footer {
-  text-align: center;
-}
-
-.title {
-  margin-bottom: 20px;
-}
 
 .post-card {
   text-align: left;
