+++ src/pages/pages.css
@@ -1,16 +1,24 @@
 .postlist {
-  width: 40%;
-  margin: auto;
+  max-width: 600px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
   text-align: center;
+  position: relative;
 }
 
 .post {
-  width: 40%;
-  margin: auto;
+  max-width: 600px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+  font-size: 2rem;
 }
 
 .page-content {
-  width: 40%;
   margin: auto;
   text-align: center;
 }
